Changes: - Moved the Theme, Navigation related things to App folder.
