+++ composeApp/src/androidMain/kotlin/omkar/android/projects/MyApplication.kt
@@ -1,7 +1,7 @@
 package omkar.android.projects
 
 import android.app.Application
-import omkar.android.projects.di.commonModule
+import omkar.android.projects.app.di.commonModule
 import org.koin.core.context.startKoin
 
 class MyApplication : Application() {
