+++ composeApp/src/commonMain/kotlin/omkar/android/projects/App.kt
@@ -2,8 +2,8 @@ package omkar.android.projects
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.*
-import omkar.android.projects.presentation.theme.AppTheme
-import omkar.android.projects.presentation.navigation.MyAppNavigation
+import omkar.android.projects.app.theme.AppTheme
+import omkar.android.projects.app.navigation.MyAppNavigation
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
