+++ composeApp/src/commonMain/kotlin/omkar/android/projects/app/components/TextFields.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.sp
-import omkar.android.projects.presentation.theme.LocalAppColors
+import omkar.android.projects.app.theme.LocalAppColors
 
 /**
  - Configuration class for text styling
