+++ composeApp/src/commonMain/kotlin/omkar/android/projects/app/theme/AppTheme.kt
@@ -1,4 +1,4 @@
-package omkar.android.projects.presentation.theme
+package omkar.android.projects.app.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
