+++ composeApp/src/commonMain/kotlin/omkar/android/projects/app/widget/SearchBarWidget.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import omkar.android.projects.app.components.percentOfScreenWidth
-import omkar.android.projects.presentation.theme.ThemeColors
+import omkar.android.projects.app.theme.ThemeColors
 
 @Composable
 fun SearchBarWidget(
