+++ composeApp/src/commonMain/kotlin/omkar/android/projects/app/widget/TopBarWidget.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
 import omkar.android.projects.app.components.percentOfScreenHeight
-import omkar.android.projects.presentation.theme.LocalAppColors
+import omkar.android.projects.app.theme.LocalAppColors
 
 @Composable
 fun TopBarWidget(
